+++ account/src/main/java/com/kimtaehoonki/account/presentation/MemberController.java
@@ -1,14 +1,15 @@
-package com.kimtaehoonki.account.application;
+package com.kimtaehoonki.account.presentation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 @Controller
 public class MemberController {
     @PostMapping("/auth/login")
-    public void login() {
+    public void login(@RequestBody LoginRequestDto dto) {
 
     }
