+++ account/src/main/java/com/kimtaehoonki/account/presentation/dto/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.kimtaehoonki.account.presentation.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginRequestDto {
+    private String userid;
+    private String password;
+}
